moving provider tests to providers
