+++ nbviewer/providers/gist/tests/test_gist.py
@@ -8,7 +8,7 @@
 
 import requests
 
-from .base import NBViewerTestCase, FormatHTMLMixin
+from ....tests.base import NBViewerTestCase, FormatHTMLMixin
 
 class GistTestCase(NBViewerTestCase):
     def test_gist(self):
