+++ nbviewer/providers/github/tests/test_github.py
@@ -10,7 +10,7 @@ import requests
 
 from unittest import SkipTest
 
-from .base import NBViewerTestCase, FormatHTMLMixin
+from ....tests.base import NBViewerTestCase, FormatHTMLMixin
 
 class GitHubTestCase(NBViewerTestCase):
     def ipython_example(self, *parts, **kwargs):
@@ -54,14 +54,14 @@ class GitHubTestCase(NBViewerTestCase):
         url = self.ipython_example('Index.ipynb', ref='rel-2.0.0')
         r = requests.get(url)
         self.assertEqual(r.status_code, 200)
-    
+
     def test_github_commit(self):
         url = self.ipython_example('Index.ipynb',
             ref='7f5cbd622058396f1f33c4b26c8d205a8dd26d16'
         )
         r = requests.get(url)
         self.assertEqual(r.status_code, 200)
-    
+
     def test_github_blob_redirect(self):
         url = self.url(
             'urls/github.com/ipython/ipython/blob/rel-2.0.0/examples',
@@ -71,7 +71,7 @@ class GitHubTestCase(NBViewerTestCase):
         self.assertEqual(r.status_code, 200)
         # verify redirect
         self.assertIn('/github/ipython/ipython/blob/master', r.request.url)
-    
+
     def test_github_raw_redirect(self):
         url = self.url(
             'urls/raw.github.com/ipython/ipython/rel-2.0.0/examples',
@@ -81,7 +81,7 @@ class GitHubTestCase(NBViewerTestCase):
         self.assertEqual(r.status_code, 200)
         # verify redirect
         self.assertIn('/github/ipython/ipython/blob/rel-2.0.0/examples', r.request.url)
-    
+
 
     def test_github_rawusercontent_redirect(self):
         """Test GitHub's new raw domain"""
@@ -93,7 +93,7 @@ class GitHubTestCase(NBViewerTestCase):
         self.assertEqual(r.status_code, 200)
         # verify redirect
         self.assertIn('/github/ipython/ipython/blob/rel-2.0.0/examples', r.request.url)
-    
+
     def test_github_raw_redirect_2(self):
         """test /url/github.com/u/r/raw/ redirects"""
         url = self.url(
@@ -104,7 +104,7 @@ class GitHubTestCase(NBViewerTestCase):
         self.assertEqual(r.status_code, 200)
         # verify redirect
         self.assertIn('/github/ipython/ipython/blob/rel-2.0.0', r.request.url)
-    
+
     def test_github_repo_redirect(self):
         url = self.url("github/ipython/ipython")
         r = requests.get(url)
